+++ src/main/java/org/opentripplanner/index/IndexGraphQLSchema.java
@@ -26,6 +26,7 @@ import org.opentripplanner.model.Agency;
 import org.opentripplanner.model.FeedScopedId;
 import org.opentripplanner.model.Route;
 import org.opentripplanner.model.Stop;
+import org.opentripplanner.model.StopTimeKey;
 import org.opentripplanner.model.Trip;
 import org.opentripplanner.model.calendar.ServiceDate;
 import org.opentripplanner.routing.edgetype.SimpleTransfer;
@@ -34,6 +35,7 @@ import org.opentripplanner.routing.graph.GraphIndex;
 import org.opentripplanner.routing.trippattern.RealTimeState;
 import org.opentripplanner.routing.vertextype.TransitVertex;
 import org.opentripplanner.updater.GtfsRealtimeFuzzyTripMatcher;
+import org.rutebanken.netex.model.Notice;
 
 import java.text.ParseException;
 import java.util.ArrayList;
@@ -88,6 +90,8 @@ public class IndexGraphQLSchema {
 
     public GraphQLOutputType clusterType = new GraphQLTypeReference("Cluster");
 
+    public GraphQLOutputType noticeType = new GraphQLTypeReference("Notice");
+
     public GraphQLOutputType patternType = new GraphQLTypeReference("Pattern");
 
     public GraphQLOutputType routeType = new GraphQLTypeReference("Route");
@@ -124,6 +128,9 @@ public class IndexGraphQLSchema {
             }
             if (o instanceof Agency){
                 return (GraphQLObjectType) agencyType;
+            }
+            if (o instanceof Notice) {
+                return (GraphQLObjectType) noticeType;
             }
             return null;
         }
@@ -429,7 +436,41 @@ public class IndexGraphQLSchema {
               	.type(Scalars.GraphQLString)
               	.dataFetcher(environment -> ((TripTimeShort) environment.getSource()).headsign)
               	.build())
+            .field(GraphQLFieldDefinition.newFieldDefinition()
+                    .name("notices")
+                    .type(new GraphQLList(noticeType))
+                    .argument(GraphQLArgument.newArgument()
+                            .name("gtfsId")
+                            .type(Scalars.GraphQLString)
+                            .build())
+                    .dataFetcher(environment -> {
+                        TripTimeShort tts = (TripTimeShort) environment.getSource();
+                        return index.getNoticesByEntity(new StopTimeKey(tts.tripId, tts.stopIndex));
+                    })
+                    .build())
             .build();
+
+        noticeType = GraphQLObjectType.newObject()
+                .name("Notice")
+                .field(GraphQLFieldDefinition.newFieldDefinition()
+                        .name("Id")
+                        .type(Scalars.GraphQLString)
+                        .dataFetcher(
+                                environment -> ((Notice) environment.getSource()).getId())
+                        .build())
+                .field(GraphQLFieldDefinition.newFieldDefinition()
+                        .name("Text")
+                        .type(Scalars.GraphQLString)
+                        .dataFetcher(
+                                environment -> ((Notice) environment.getSource()).getText())
+                        .build())
+                .field(GraphQLFieldDefinition.newFieldDefinition()
+                        .name("PublicCode")
+                        .type(Scalars.GraphQLString)
+                        .dataFetcher(
+                                environment -> ((Notice) environment.getSource()).getPublicCode())
+                        .build())
+                .build();
 
         tripType = GraphQLObjectType.newObject()
             .name("Trip")
@@ -538,6 +579,17 @@ public class IndexGraphQLSchema {
                 })
                 .build())
             .field(GraphQLFieldDefinition.newFieldDefinition()
+                    .name("notices")
+                    .type(new GraphQLList(noticeType))
+                    .argument(GraphQLArgument.newArgument()
+                            .name("gtfsId")
+                            .type(Scalars.GraphQLString)
+                            .build())
+                    .dataFetcher(environment -> {
+                        return index.getNoticesByEntity((Trip) environment.getSource());
+                    })
+                    .build())
+            .field(GraphQLFieldDefinition.newFieldDefinition()
                 .name("geometry")
                 .type(Scalars.GraphQLString) //TODO: Should be geometry
                 .dataFetcher(environment -> index.patternForTrip
@@ -568,7 +620,7 @@ public class IndexGraphQLSchema {
                 .name("id")
                 .type(new GraphQLNonNull(Scalars.GraphQLID))
                 .dataFetcher(environment -> relay.toGlobalId(
-                    patternType.getName(), ((TripPattern) environment.getSource()).code))
+                    patternType.getName(), ((TripPattern) environment.getSource()).getCode()))
                 .build())
             .field(GraphQLFieldDefinition.newFieldDefinition()
                 .name("route")
@@ -588,7 +640,7 @@ public class IndexGraphQLSchema {
             .field(GraphQLFieldDefinition.newFieldDefinition()
                 .name("code")
                 .type(new GraphQLNonNull(Scalars.GraphQLString))
-                .dataFetcher(environment -> ((TripPattern) environment.getSource()).code)
+                .dataFetcher(environment -> ((TripPattern) environment.getSource()).getCode())
                 .build())
             .field(GraphQLFieldDefinition.newFieldDefinition()
                 .name("headsign")
@@ -622,6 +674,17 @@ public class IndexGraphQLSchema {
                 .dataFetcher(environment ->
                     ((TripPattern) environment.getSource()).semanticHashString(null))
                 .build())
+            .field(GraphQLFieldDefinition.newFieldDefinition()
+                    .name("notices")
+                    .type(new GraphQLList(noticeType))
+                    .argument(GraphQLArgument.newArgument()
+                            .name("gtfsId")
+                            .type(Scalars.GraphQLString)
+                            .build())
+                    .dataFetcher(environment -> {
+                        return index.getNoticesByEntity((TripPattern) environment.getSource());
+                    })
+                    .build())
             .build();
 
 
@@ -784,6 +847,11 @@ public class IndexGraphQLSchema {
                 }
                 return null;
             }))
+            .field(GraphQLFieldDefinition.newFieldDefinition()
+                    .name("notices")
+                    .type(new GraphQLList(noticeType))
+                    .dataFetcher(environment -> index.graph.getNoticesByElement().values())
+                    .build())
             .field(GraphQLFieldDefinition.newFieldDefinition()
                 .name("agencies")
                 .description("Get all agencies for the specified graph")