+++ src/main/java/org/opentripplanner/model/Transfer.java
@@ -119,7 +119,7 @@ public final class Transfer implements Serializable {
                 + ">";
     }
 
-    private static String toStrOpt(String lbl, IdentityBean arg1, IdentityBean arg2) {
+    private static String toStrOpt(String lbl, TransitEntity arg1, TransitEntity arg2) {
         return  (arg1 == null ? "" : (lbl + arg1.getId() + ".." + arg2.getId()));
     }
 }